WIP: added test to parse FitCSV and then write file back to disk and compare
This test is largely working but it's slow and there's currently a bug where rows (in the original) that have a empty fields are not being respected. In the original, many lines have lots of commas at the end of the line to represent these empty fields. The generated file does not have these. WIll fix tomorrow.