Added MultipleValueAggregator
the idea here is that we don't want to loop over every record in an activity in order to calculate things like min, max, or averages. We want to calculate these as we go. To do this, we have an aggregator object that can keep running totals as needed. The MultipleValueAggregator will create multiple aggregators for each of the fields (keys) provided. For example, we can give the MVA a "record" array on each loop through the records and ask it to keep track of max HR, max speed, and avg cadence all at once. The downside is that we will calculate many more things than we need. We'll have to do performance testing to see if this is really better than multiple passes through the records array.