Minor changes to GlobalProfileGenerator
Moved generate profile functionality out of the TestCommand and into its own command `GenerateProfileCommand` and added ability to specify an output path. Updated the original python file to address whitespace issues that are not really relevant. In the future, a linter would take care of this. Also temporarily changed the class of the generated profile to be GlobalProfile2.