Changed `mesg_num` to `global_number` in GlobalProfile and generator
Part of an ongoing battle to come up with names that result in the least amount of confusion. The FIT SDK documentation is inconsistent. This value is often referred to as the message type or the global message type, but message type is also used to differentiate between data and definition messages. Message type is also used in the documentation to refer to what we call message.name. `Message.global_number` seems most appropriate.