Added 'size' to BASE_TYPES
Originally wanted to convert BASE_TYPES to an array of FIT\BASE_TYPE objects, but you can't declare a constant to be an object, so I'll leave the GlobalProfile constants as arrays of arrays and let the accessor class do the conversion to objects. Also deleted profile.php.txt - originally thought of as the default place to land a new auto-generated GlobalProfile, but there's no need to keep track of what is effectively just a second copy of GlobalProfile.