Updates to address new understanding of multi-value fields
Fields can contain multiple values. Thus the size of a field can be any multiple of the size of the base type for that field. The problem is the FITCSVParser doesn't know about base type yet, so it stores the number of values and then the FITCSVWriter uses that when it writes a CSV file. So we're currently in a states where the CSVWriter needs to behave differently depending on whether the FIT Messages were created from a .fit or a .csv file. This isn't good and needs to be fixed. In the meantime, there are some hacks in place to handle it for now.