FITCSVParser now gets FieldDefinitions from the GlobalProfile
added code to FITCSVParser to make message building more inline with that of FITParser. Fields are now a non-associative array, and the FieldDefinitions are filled out from the GlobalProfile. FieldDefinition gained a `initFromGlobalProfileByNames($message_type_name, $properties) : FieldDefinition` method to facilitate. The testing of FITCSVParser is very brittle and needs work.