Added BaseType and FieldType classes
Updated FieldDefinition to be able to set FieldDefinition.type if passed an array of properties or a constructed BaseType or FieldType. Also fixed a bug - setComponents and setSubfields would not have worked if passed constructed objects - only worked on properties array. Much of this work needs to be abstracted and moved into a trait.