Removing `$raw_value` from FIT\Field class
Fields (data) were storing the value in both $value and $raw_value. This may be helpful down the road, but we don't need or use it now. FieldDefinitions were storing `number_of_values` in raw_value before we understood what this number really meant. That is no longer needed and bring FITParser and FITCSVParser more inline with each other.