updated FIT\DataMessage::getValidFields() to work with fields that don't have a def_num
Fields that are created from the FITCSVParser do no have a corresponding definition field in the case of subfields or expanded components. While it may be possible to reverse engineer this, that's not the current plan. As a result, sparsely populated fields (ie: no def_num and thus no FieldDefinition) need to be evaluated slightly differently. I don't love this implementation. Should be revisited.